closure

+++ application/tool/controller/Detailstatic.php
@@ -141,6 +141,11 @@ class Detailstatic extends Common
                         'next_article' => $next_article
                     ]
                 );
+                //判断模板是否存在
+                if (!file_exists('article')) {
+                    $this->make_error("article");
+                    die;
+                }
                 $make_web = file_put_contents('article/article' . $item["id"] . '.html', $content);
                 //开始同步数据库
                 if ($make_web) {
@@ -229,7 +234,11 @@ class Detailstatic extends Common
                         'next_article' => $next_article
                     ]
                 );
-
+                //判断模板是否存在
+                if (!file_exists('news')) {
+                    $this->make_error("news");
+                    die;
+                }
                 $make_web = file_put_contents('news/news' . $item["id"] . '.html', $content);
                 //开始同步数据库
                 if ($make_web) {
@@ -247,9 +256,6 @@ class Detailstatic extends Common
                         $articleCountModel->save();
                     }
                     $limit = $item["id"];
-                }else {
-                    $this->make_error("news");
-                    exit();
                 }
             }
         }
